requirement.txt
pandas
python-dotenv
pymongo[srv]
tushare